Add trend analytics tab with weekly sparkline
