+++ app/src/main/java/com/concepts_and_quizzes/cds/ui/english/pyqp/PyqAnalyticsScreen.kt
@@ -1,33 +1,42 @@
 package com.concepts_and_quizzes.cds.ui.english.pyqp
 
 import androidx.compose.foundation.Canvas
-import androidx.compose.animation.core.Animatable
-import androidx.compose.animation.core.tween
-import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.material3.*
-import androidx.compose.runtime.*
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.material3.Button
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.FilterChip
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Switch
+import androidx.compose.material3.Tab
+import androidx.compose.material3.TabRow
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.geometry.Size
-import androidx.compose.ui.graphics.Path
-import androidx.compose.ui.graphics.PathEffect
-import androidx.compose.ui.graphics.drawscope.Stroke
-import androidx.compose.ui.graphics.drawscope.clipRect
+import androidx.compose.ui.semantics.contentDescription
+import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavController
 import com.concepts_and_quizzes.cds.data.analytics.db.TopicStat
-import com.concepts_and_quizzes.cds.data.analytics.db.TrendPoint
 import com.concepts_and_quizzes.cds.data.analytics.repo.AnalyticsRepository
-import androidx.compose.ui.semantics.contentDescription
-import androidx.compose.ui.semantics.semantics
-import androidx.compose.ui.platform.LocalContext
-import android.provider.Settings
-import java.time.Instant
-import java.time.LocalDate
-import java.time.ZoneId
+import com.concepts_and_quizzes.cds.ui.analytics.TrendTab
+import androidx.compose.material3.contentColorFor
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -35,7 +44,6 @@ fun PyqAnalyticsScreen(
     nav: NavController,
     vm: PyqAnalyticsViewModel = hiltViewModel()
 ) {
-    val window by vm.window.collectAsState()
     val stats by vm.stats.collectAsState()
     val trend by vm.trend.collectAsState()
     val tab by vm.tab.collectAsState()
@@ -43,7 +51,7 @@ fun PyqAnalyticsScreen(
 
     Scaffold(topBar = { TopAppBar(title = { Text("PYQ Analytics") }) }) { pad ->
         Column(Modifier.padding(pad).padding(16.dp)) {
-            FilterChipRow(window) { vm.setWindow(it) }
+            FilterChipRow(vm)
 
             Row(verticalAlignment = Alignment.CenterVertically) {
                 Text("High contrast")
@@ -97,15 +105,13 @@ fun PyqAnalyticsScreen(
 }
 
 @Composable
-private fun FilterChipRow(
-    selected: AnalyticsRepository.Window,
-    onSelect: (AnalyticsRepository.Window) -> Unit
-) {
+private fun FilterChipRow(vm: PyqAnalyticsViewModel) {
+    val selected by vm.window.collectAsState()
     Row(horizontalArrangement = Arrangement.spacedBy(8.dp)) {
         AnalyticsRepository.Window.entries.forEach { w ->
             FilterChip(
                 selected = w == selected,
-                onClick = { onSelect(w) },
+                onClick = { vm.setWindow(w) },
                 label = {
                     Text(
                         when (w) {
@@ -161,75 +167,6 @@ private fun TopicBarList(stats: List<TopicStat>, highContrast: Boolean) {
                 Spacer(Modifier.width(8.dp))
                 Text("$pct %", color = content)
             }
-        }
-    }
-}
-
-@Composable
-private fun TrendTab(points: List<TrendPoint>, highContrast: Boolean) {
-    if (points.isEmpty()) {
-        Text("Attempt at least one paper to see your trend.", Modifier.padding(24.dp))
-        return
-    }
-
-    SparkLineChart(points, highContrast)
-    Row(horizontalArrangement = Arrangement.SpaceBetween, modifier = Modifier.fillMaxWidth()) {
-        points.forEach {
-            val pct = "%.0f".format(it.percent)
-            Text(pct, style = MaterialTheme.typography.labelSmall)
-        }
-    }
-}
-
-@Composable
-private fun SparkLineChart(points: List<TrendPoint>, highContrast: Boolean) {
-    val max = points.maxOf { it.percent }
-    val anim = remember { Animatable(0f) }
-    val context = LocalContext.current
-    val animationsDisabled = remember {
-        Settings.Global.getFloat(
-            context.contentResolver,
-            Settings.Global.ANIMATOR_DURATION_SCALE,
-            1f
-        ) == 0f
-    }
-    LaunchedEffect(points, animationsDisabled) {
-        if (animationsDisabled) {
-            anim.snapTo(1f)
-        } else {
-            anim.animateTo(1f, tween(600))
-        }
-    }
-
-    val desc = points.joinToString {
-        val pct = "%.0f".format(it.percent)
-        val week = Instant.ofEpochMilli(it.weekStart).atZone(ZoneId.systemDefault()).toLocalDate()
-        "$week : $pct percent"
-    }
-
-    val baseColor = if (highContrast) MaterialTheme.colorScheme.onBackground else MaterialTheme.colorScheme.primary
-    val effect = if (highContrast) PathEffect.dashPathEffect(floatArrayOf(10f, 10f)) else null
-
-    Canvas(
-        Modifier
-            .fillMaxWidth()
-            .height(120.dp)
-            .padding(8.dp)
-            .semantics { contentDescription = desc }
-    ) {
-        val stepX = if (points.size == 1) 0f else size.width / (points.size - 1)
-        val path = Path()
-        points.forEachIndexed { i, p ->
-            val x = i * stepX
-            val y = size.height * (1 - p.percent / max.coerceAtLeast(1f))
-            if (i == 0) path.moveTo(x, y) else path.lineTo(x, y)
-        }
-        clipRect(right = size.width * anim.value) {
-            drawPath(
-                path = path,
-                color = baseColor,
-                style = Stroke(width = 4.dp.toPx(), pathEffect = effect)
-            )
         }
     }
 }
